Move satisfies up to Container, remove containerContext tests
